# etc/docs/swagger.yaml
definitions:
...
  dto.HTTPErrorResp:
    properties:
      metadata:
        $ref: '#/definitions/dto.Meta'
    type: object
...
  dto.HttpSuccessResp:
    properties:
      data:
        x-order: "1"
      metadata:
        allOf:
        - $ref: '#/definitions/dto.Meta'
        x-order: "0"
      pagination:
        allOf:
        - $ref: '#/definitions/dto.Pagination'
        x-order: "2"
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.Meta:
    properties:
      error:
        type: object
        x-order: "4"
      message:
        type: string
        x-order: "3"
      path:
        type: string
        x-order: "0"
      status:
        type: string
        x-order: "2"
      status_code:
        type: integer
        x-order: "1"
      timestamp:
        type: string
        x-order: "5"
    type: object
  dto.Pagination:
    properties:
      current_elements:
        type: integer
        x-order: "1"
      current_page:
        type: integer
        x-order: "0"
      cursor_end:
        type: string
        x-order: "7"
      cursor_start:
        type: string
        x-order: "6"
      sort_by:
        type: string
        x-order: "4"
      sort_dir:
        type: string
        x-order: "5"
      total_elements:
        type: integer
        x-order: "3"
      total_pages:
        type: integer
        x-order: "2"
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RegisterRequest:
    properties:
      age:
        maximum: 150
        minimum: 1
        type: integer
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/entity.Role'
        enum:
        - admin
        - user
        - guest
    required:
    - age
    - email
    - name
    - password
    - role
    type: object
  dto.TokenResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: integer
      refreshToken:
        type: string
    type: object
...
  entity.Role:
    enum:
    - admin
    - user
    - guest
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
    - RoleGuest
  entity.User:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      updated_at:
        type: string
    type: object
host: localhost:8181
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Clean Architecture CRUD API with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go-Far
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access/refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.HttpSuccessResp'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
      summary: Login user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate a new access token using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.HttpSuccessResp'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.HttpSuccessResp'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
- https
swagger: "2.0"